test(type): improve & document
