+++ app/src/test/java/com/neo/speaktouch/view/CustomList.kt
@@ -1,0 +1,41 @@
+/*
+ * An example of a custom list.
+ *
+ * Copyright (C) 2023 Irineu A. Silva.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.neo.speaktouch.view
+
+import android.content.Context
+import android.view.View
+import android.view.accessibility.AccessibilityNodeInfo
+
+class CustomList(
+    context: Context
+) : View(context) {
+
+    override fun onInitializeAccessibilityNodeInfo(
+        info: AccessibilityNodeInfo
+    ) {
+        super.onInitializeAccessibilityNodeInfo(info)
+
+        info.collectionInfo = AccessibilityNodeInfo.CollectionInfo(
+            1,
+            1,
+            false
+        )
+    }
+}
+
